openvino: split quantized weight graph builders
