+++ ggml/src/ggml-openvino/ggml-quants.h
@@ -53,6 +53,9 @@ void extract_q6_k_data(const ggml_tensor * tensor,
 void extract_mxfp4_data(const ggml_tensor * tensor, ov::Tensor & weights_arr, ov::Tensor & scales_arr);
 
 static constexpr size_t GGML_QUANTIZATION_GROUP_SIZE = 32;
+static constexpr size_t MXFP4_BLOCK_SIZE = 32;
+static constexpr size_t MXFP4_BLOCK_QS_SIZE = MXFP4_BLOCK_SIZE / 2;
+static constexpr size_t MXFP4_BLOCK_BYTES = sizeof(uint8_t) + MXFP4_BLOCK_QS_SIZE;
 
 // If for_gather_matmul is true, the weight tensor may be N-D (e.g. 3D MoE expert weights
 // [n_expert, rows, cols]). The dequantization chain (Convert->[Subtract]->Multiply) is built as
